feat: implement timeout in HTTP and MQTT notifications

+++ subscription-service/src/main/kotlin/com/egm/stellio/subscription/utils/DBConversionUtils.kt
@@ -21,5 +21,5 @@ fun toJsonString(entry: Any?): String? = (entry as? Json)?.asString()
 inline fun <reified T : Enum<T>> toEnum(entry: Any) = enumValueOf<T>(entry as String)
 inline fun <reified T : Enum<T>> toOptionalEnum(entry: Any?) =
     (entry as? String)?.let { enumValueOf<T>(it) }
-fun toInt(entry: Any?): Int = (entry as Long).toInt()
+fun toInt(entry: Any?): Int = entry as Int
 fun toNullableInt(entry: Any?): Int? = entry as? Int
